Bump version to 0.1.3; CTR & API updates
Release/version bump to 0.1.3 across headers and Windows resource files. Add a ctr_block parameter to cipher callback signatures (include/CryptoForge/cf_cipher.h) and propagate through cf_cipher and all cipher wrappers, making CTR mode caller-managed (counter is non-const). Update AES CTR implementation/header signatures to accept a mutable counter and adjust callers to pass opts->ctr_block. Miscellaneous fixes: suppress cast-qual warning in AEAD decrypt tag path, correct CF_Enc_GetName signature, tighten casts in AES core and base64, and add stricter compiler/compile-option handling in CMake (and clean up CMakeLists). Remove ENABLE_CF_TESTS_VERBOSE from VS Code tasks.json and tidy tests CMakeLists.

+++ src/crypto/cipher/aes/aes_core.c
@@ -66,7 +66,7 @@ static const uint8_t rCon[11] = {
 static FORCE_INLINE uint8_t XtimeByte(uint8_t b) {
     // Create mask: 0x1b if high bit is set, 0x00 otherwise
     uint8_t mask = -(b >> 7);  // high bit: 1 -> 0xFF, 0 -> 0x00
-    return (b << 1) ^ (0x1b & mask);
+    return (uint8_t)((b << 1) ^ (0x1b & mask));
 }
 
 /*
